1sec

+++ saveDelete.jsp
@@ -7,11 +7,12 @@ String basePath = request.getScheme()+"://"+request.getServerName()+":"+request.
 <%@ include file="db.jsp"%><%
 Statement sts=conn.createStatement();
 int sno = Integer.parseInt(request.getParameter("sno"));
-String sql="delete from save where s_no="+sno+"";
+int uid = (int)session.getAttribute("uid");
+String sql="delete from save where s_no="+sno+" and s_owner="+uid;
 if (sts.executeUpdate(sql) == 1) {
 %>
 <p>删除成功</p>
-<% session.sendRedirect("saveQuery.jsp"); %>
+<% response.sendRedirect("saveQuery.jsp"); %>
 <% } else { %>
 <p>删除失败</p>
 <% } %>
